DatasetView: make getChannelCount() return int
There are many reasons why ImageJ2 does not support more than 2 Gi
channels at the moment, so having the API return long here is
disingenuous. And it is more convenient for it to return int, to avoid
casts in various places.

+++ core/data/src/main/java/imagej/data/display/DatasetView.java
@@ -64,7 +64,7 @@ public interface DatasetView extends DataView {
 
 	CompositeXYProjector<? extends RealType<?>> getProjector();
 
-	long getChannelCount();
+	int getChannelCount();
 
 	/**
 	 * Gets the minimum value in the <em>display</em> range, for the given
